+++ jest.config.js
@@ -1,6 +1,6 @@
 module.exports = {
   testEnvironment: "jsdom",
   testMatch: [
-    "**/src/components/**/**/*.test.[jt]s?(x)",
+    "**/src/components/**/**/*.test.tsx",
   ],
 };