+++ src/components/base/IconButton/IconButton.test.tsx
@@ -103,7 +103,7 @@ describe("<IconButton />", () => {
     expect(onClick).toHaveBeenCalledTimes(1);
   });
 
-  it("should render component disabled", async () => {
+  it("should render component disabled", () => {
     const onClick = jest.fn();
 
     render(